requirements.txt
dj-database-url==0.4.1
Django==1.10.4
django-pyodbc-azure==1.10.4.0
djangorestframework==3.6.2
gunicorn==19.6.0
psycopg2==2.6.2
pyodbc==4.0.14
whitenoise==3.2
